+++ packages/lib/server/service/InsightsBookingBaseService.ts
