+++ packages/lib/server/service/__tests__/insightsBooking.integration-test.ts
@@ -5,7 +5,7 @@ import { describe, expect, it } from "vitest";
 import prisma from "@calcom/prisma";
 import { BookingStatus, MembershipRole } from "@calcom/prisma/enums";
 
-import { InsightsBookingBaseService } from "../insightsBookingBase";
+import { InsightsBookingBaseService } from "../InsightsBookingBaseService";
 
 const NOTHING_CONDITION = Prisma.sql`1=0`;
